Add Passport jwt in CommentController

+++ controllers/CommentController.js
@@ -1,4 +1,5 @@
-const models = require("../models")
+const models = require("../models");
+const e = require("express");
 
 const response = {
     message: "Your Message",
@@ -9,7 +10,9 @@ const response = {
 class Controller {
     static async create(req, res) {
         try {
-            const data = await models.Comment.create(req.body)
+            let obj = req.body;
+            obj.author_id = req.user.id;
+            const data = await models.Comment.create(obj)
             response.data = data;
             response.message = "Data is successfully created";
         
@@ -42,16 +45,21 @@ class Controller {
 
     static async find(req, res) {
         try {
-            const data =  await models.Comment.findByPk(req.params.id, {
-                include: [
-                    { model: models.Author },
-                    { model: models.Post }
-                ]
-            })
+            const data = await models.Comment.findOne({
+                where: { id: req.params.id, author_id: req.user.id }
+            });
+            if (data == null) {
+                response.status = "Fail"
+                response.message = "Data doesn't exist";
+                response.data = data;
+
+                return res.json(response)
+            }
+            response.status = "Success"
             response.data = data;
-            response.message = "Data successfully retrieved";
-    
-            res.json(response)
+            response.message = "Data is successfully retrieved";
+        
+            res.status(200).json(response);
         } catch (error) {
             response.status = "Fail",
             response.message = error.message;
@@ -61,15 +69,27 @@ class Controller {
 
     static async update(req, res) {
         try {
-            await models.Comment.update(req.body, {
-                where: {
-                    id: req.params.id,
-                },
-            })
-            response.data = { Comment_Id : req.params.id }
-            response.message = "Data is successfully updated";
-        
-            res.status(201).json(response);
+            const data = await models.Comment.findOne({
+                where: { id: req.params.id, author_id: req.user.id }
+            });
+            if (data == null) {
+                response.status = "Fail"
+                response.message = "Data doesn't exist";
+                response.data = data;
+
+                return res.status(401).json(response)
+            } else {
+                let obj = req.body;
+                obj.author_id = req.user.id
+                await models.Comment.update(obj, {
+                    where: {
+                        id: req.params.id,
+                    },
+                })
+                response.status = "Success"
+                response.message = "Data is successfully updated";
+                res.status(201).json(response);
+            }
         } catch (error) {
             response.status = "Fail",
             response.message = error.message,
@@ -79,15 +99,26 @@ class Controller {
 
     static async delete(req, res) {
         try {
-            await models.Comment.destroy({
-                where: {
-                    id: req.params.id,
-                }   
-            })
-            response.data = { id : req.params.id }
-            response.message = "Data is successfully deleted";
-        
-            res.status(201).json(response);
+            const data = await models.Comment.findOne({
+                where: { id: req.params.id, author_id: req.user.id }
+            });
+            if (data == null) {
+                response.status = "Fail"
+                response.message = "Data doesn't exist";
+                response.data = data;
+
+                return res.status(401).json(response)
+            } else {
+                await models.Comment.destroy({
+                    where: {
+                        id: req.params.id,
+                    }   
+                })
+                response.data = { id : req.params.id }
+                response.message = "Data is successfully deleted";
+            
+                res.status(201).json(response);
+            }
         } catch (error) {
             response.status = "Fail",
             response.message = error.message,
